Created player prefabs and cameraPivot additions
CameraPivot is a pivot point the new camera will be using, but right now it seems to acting a little weird. There is a camera put on the CameraPivot, but it is not what the new camera will be. I just wanted to see its view point. If you wanna see it hop into my scene.