Camera Changes
Major changes to the camera, functions a little differently now. To keep it mostly simple, the camera will keep up with the player in perfect pace, no shakiness. However, the player cannot walk around the camera, if the player moves to the right the camera also moves to the right. (Haven't figured out a reliable way to change that yet). You can adjust the camera by moving your mouse around though. You can also change its zoom, but it has trouble detecting the scrollwheel input. Camera controls currently do not exist for controllers, I want to set it up with that, but I lack the ability to use a controller to figure that out.
Made the camera a prefab and cameraPivot a prefab. If the camera does not have a cameraPivot to work with it will revert to the old camera code.