Camera Adjustments
There was an issue with the camera where it wouldn't fully lock on to the player. This fixes that issue, as well as making the transition when switching players better (I forgot to test this like an idiot).
Also, I added a collision check that stops the camera from passing through objects, but unfortunately the camera gets stuck when using this collision check logic. I commented out the code that specifcally does the collision check in CameraScript, uncomment it if you want to see the issue I was running into.